requirements.txt
Django==1.11.8
mysqlclient==1.3.12
configparser==3.5.0
Pillow==5.0.0
django-simple-captcha==0.5.6
requests==2.18.4 
celery==3.1.25
django-formtools==2.1
django-bootstrap3==9.1.0
itsdangerous==0.24
django-datatables-view==1.14.0
djangorestframework-bulk==0.2.1
djangorestframework==3.7.7
redis==2.10.6
django-simple-history==1.9.0
django-model-utils==3.1.1
django-celery==3.2.2
celery-with-redis==3.0
pyecharts==0.4.1